backend/requirements.txt
fastapi
gunicorn
uvicorn
psycopg2-binary
pyopenssl
pubchempy
requests
dotenv
numpy
matplotlib
SQLAlchemy>=2.0
Werkzeug>=2.0
ripgrep
